requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorlog==6.8.2
flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
SQLAlchemy==2.0.25
twilio==8.13.0
typing-extensions==4.9.0
urllib3==2.2.0
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
